all: refactor to use composition and simplify state
Compose ReaderCloser from Reader, WriterCloser from Writer
to reuse a bunch of duplicated code and also fix up the
state management to avoid data races. Employing once
atomics and a RWMutex to avoid sending data on closed
channels since the main consumer of this project
drive makes asynchronous requests to each of the
objects and we've seen send on closed channels in the wild.

This commit is a long awaited fix-up that I hadn't
implemented due to lack of time, but alas the chickens
have come home to roost.